setting up framework for fine-grained authorization and filtering
Expand the framework so we can do fine-grained authorization (for instance
at individual API call level). There is potential for using dynamic context
such as from OAuth in addition to configuration driven context such as from
an Agent or App. It is hoped that this proposal will pave way for
implementation of authorization engines which can be registered and used with
the framework.

The example `examples/basic/mcp_basic_agent_with_auth` hopefully illustrates
all aspects of this extension.

Signed-off-by: smuralidharan <smuralidharan@manetu.com>